--- modulo4/intro-express/template-exercicios-express/src/usuarios.ts
@@ -1,38 +1,0 @@
-type Usuarios = {
-    id: number,
-    name: string,
-    phone: string,
-    email: string,
-    website: string
-}
-export const usuarios: Usuarios[] =  [{
-    id: 1,
-    name: "Leanne Graham",
-    phone: "493-170-9623",
-    email: "Sincere@april.biz",
-    website: "kale.biz",
-}, {
-    id: 2,
-    name: "Diego Lima",
-    phone: "493-854-1254",
-    email: "newdev@april.biz",
-    website: "frontend-dev.biz",
-},{
-    id: 3,
-    name: "Murilo Terenciani",
-    phone: "589-458-9623",
-    email: "monitor@labenu.com.br",
-    website: "fullstack-dev.biz",
-},{
-    id: 4,
-    name: "Eduardo",
-    phone: "041-998-5487",
-    email: "fullstack_developer@com.br",
-    website: "fullstack-dev.com.br",
-},{
-    id: 5,
-    name: "Marcela",
-    phone: "051-848-5487",
-    email: "developer@com.br",
-    website: "fullstack-dev.com.br",
-}]
